fixed money type

+++ internal/models/subscription.go
@@ -30,7 +30,7 @@ type SubscriptionDetails struct {
 	Id         uuid.UUID `json:"id,omitempty"`
 	UserID     uuid.UUID `json:"user_id,omitempty"`
 	MonthCount int64     `json:"month_count,omitempty"`
-	Money      int64     `json:"money,omitempty"`
+	Money      float32   `json:"money,omitempty"`
 }
 
 type PaymentDetails struct {
@@ -39,7 +39,7 @@ type PaymentDetails struct {
 	Id         uuid.UUID `json:"id,omitempty"`
 	UserID     uuid.UUID `json:"user_id,omitempty"`
 	MonthCount int64     `json:"month_count,omitempty"`
-	Money      int64     `json:"money,omitempty"`
+	Money      float32   `json:"money,omitempty"`
 }
 
 func (subscription *Subscription) Sanitize() {
